requirements.txt
kafka-python
tldextract
bytewax[rdkafka]
confluent_kafka